Create property package and move some classes, should be safe since using IDE built-in refactor tool
