+++ src/main/java/seedu/address/model/person/Person.java
@@ -10,6 +10,11 @@ import java.util.Set;
 
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import seedu.address.model.property.Address;
+import seedu.address.model.property.Email;
+import seedu.address.model.property.Name;
+import seedu.address.model.property.Phone;
+import seedu.address.model.property.Property;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
 
@@ -22,7 +27,7 @@ public class Person implements ReadOnlyPerson {
     private ObjectProperty<Phone> phone;
     private ObjectProperty<Email> email;
     private ObjectProperty<Address> address;
-    private HashMap<String, ObjectProperty> properties = new HashMap<>();
+    private ObjectProperty<HashMap<String, Property>> properties;
     private ObjectProperty<UniqueTagList> tags;
 
     /**
@@ -35,10 +40,12 @@ public class Person implements ReadOnlyPerson {
         this.email = new SimpleObjectProperty<>(email);
         this.address = new SimpleObjectProperty<>(address);
 
-        properties.put("n", this.name);
-        properties.put("p", this.phone);
-        properties.put("e", this.email);
-        properties.put("a", this.address);
+        HashMap<String, Property> defaultProperties = new HashMap<>();
+        defaultProperties.put("n", name);
+        defaultProperties.put("p", phone);
+        defaultProperties.put("e", email);
+        defaultProperties.put("a", address);
+        setProperties(defaultProperties);
 
         // protect internal tags from changes in the arg list
         this.tags = new SimpleObjectProperty<>(new UniqueTagList(tags));
@@ -108,6 +115,20 @@ public class Person implements ReadOnlyPerson {
         return address.get();
     }
 
+    @Override
+    public ObjectProperty<HashMap<String, Property>> properties() {
+        return properties;
+    }
+
+    @Override
+    public HashMap<String, Property> getProperties() {
+        return properties.get();
+    }
+
+    public void setProperties(HashMap<String, Property> replacement) {
+        this.properties.set(replacement);
+    }
+
     /**
      * Returns an immutable tag set, which throws {@code UnsupportedOperationException}
      * if modification is attempted.
@@ -145,5 +166,4 @@ public class Person implements ReadOnlyPerson {
     public String toString() {
         return getAsText();
     }
-
 }
