+++ src/main/java/seedu/address/model/person/ReadOnlyPerson.java
@@ -1,8 +1,14 @@
 package seedu.address.model.person;
 
+import java.util.HashMap;
 import java.util.Set;
 
 import javafx.beans.property.ObjectProperty;
+import seedu.address.model.property.Address;
+import seedu.address.model.property.Email;
+import seedu.address.model.property.Name;
+import seedu.address.model.property.Phone;
+import seedu.address.model.property.Property;
 import seedu.address.model.tag.Tag;
 import seedu.address.model.tag.UniqueTagList;
 
@@ -20,6 +26,8 @@ public interface ReadOnlyPerson {
     Email getEmail();
     ObjectProperty<Address> addressProperty();
     Address getAddress();
+    ObjectProperty<HashMap<String, Property>> properties();
+    HashMap<String, Property> getProperties();
     ObjectProperty<UniqueTagList> tagProperty();
     Set<Tag> getTags();
