+++ src/main/java/seedu/address/model/property/Property.java
@@ -1,9 +1,9 @@
-package seedu.address.model.person;
+package seedu.address.model.property;
 
 import static java.util.Objects.requireNonNull;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.person.exceptions.PropertyNotFoundException;
+import seedu.address.model.property.exceptions.PropertyNotFoundException;
 
 /**
  * A generic class that represents a property of a person. All properties of a person (including name, email, phone
