+++ src/main/java/seedu/address/model/property/exceptions/DuplicatePropertyException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.property.exceptions;
 
 /**
  * Signals that the property with the same short name already exists.
