+++ src/main/java/seedu/address/model/property/exceptions/PropertyNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.property.exceptions;
 
 /**
  * Signals that the required property has not been defined yet.
