+++ src/test/java/seedu/address/logic/commands/FindCommandTest.java
@@ -21,7 +21,7 @@ import seedu.address.model.AddressBook;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.property.NameContainsKeywordsPredicate;
 import seedu.address.model.person.ReadOnlyPerson;
 
 /**
