+++ src/test/java/seedu/address/model/person/AddressTest.java
@@ -5,6 +5,8 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import seedu.address.model.property.Address;
+
 public class AddressTest {
 
     @Test
