+++ src/test/java/seedu/address/model/person/NameTest.java
@@ -5,6 +5,8 @@ import static org.junit.Assert.assertTrue;
 
 import org.junit.Test;
 
+import seedu.address.model.property.Name;
+
 public class NameTest {
 
     @Test
