+++ src/test/java/seedu/address/model/person/PropertyTest.java
@@ -8,7 +8,8 @@ import org.junit.Test;
 import org.junit.rules.ExpectedException;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.person.exceptions.PropertyNotFoundException;
+import seedu.address.model.property.exceptions.PropertyNotFoundException;
+import seedu.address.model.property.Property;
 
 public class PropertyTest {
     @Rule
